rough infinite scroll
Infinite scroll working, cant do a ton of testing since we don't have a ton of items but it is working atm.
In debug mode its SUPER slow but on my phone it seems to be alright, we may want to unload and reload items as scrolling occurs which i sorta laid the groundwork for here.