Two small changes
When creating an item, I was able to press the submit button twice and create two duplicate items at once. I put a few lines of logic in that disable the button to prevent that.

I also just made the venmo id entry get focused when it is enabled and made an enter press trigger the updating of the id.